+++ Integra.Vision.Language/ASTNodes/Constants/BooleanNode.cs
@@ -34,7 +34,7 @@ namespace Integra.Vision.Language.ASTNodes.Constants
             this.result.Column = treeNode.Token.Location.Column;
             this.result.Line = treeNode.Token.Location.Line;
             this.result.NodeText = treeNode.Token.Text;
-            this.result.NodeType = (uint)PlanNodeTypeEnum.Constant;
+            this.result.NodeType = PlanNodeTypeEnum.Constant;
             this.result.Properties.Add("Value", treeNode.Token.Value);
             this.result.Properties.Add("DataType", treeNode.Token.Value.GetType().ToString());
         }
