+++ Integra.Vision.Language/ASTNodes/Constants/ConstantValueWithAliasNode.cs
@@ -68,7 +68,7 @@ namespace Integra.Vision.Language.ASTNodes.Constants
             this.result.Column = v.Column;
             this.result.Line = v.Line;
             this.result.NodeText = v.NodeText + " " + this.tAs + " " + a.NodeText;
-            this.result.NodeType = (uint)PlanNodeTypeEnum.ValueWithAlias;
+            this.result.NodeType = PlanNodeTypeEnum.ValueWithAlias;
             this.result.Children = new List<PlanNode>();
             this.result.Children.Add(v);
             this.result.Children.Add(a);
