+++ Integra.Vision.Language/ASTNodes/QuerySections/FromNode.cs
@@ -46,7 +46,7 @@ namespace Integra.Vision.Language.ASTNodes.QuerySections
             this.result = new PlanNode();
             this.result.Column = ChildrenNodes[0].Token.Location.Column;
             this.result.Line = ChildrenNodes[0].Token.Location.Line;
-            this.result.NodeType = (uint)PlanNodeTypeEnum.From;
+            this.result.NodeType = PlanNodeTypeEnum.From;
         }
 
         /// <summary>
