+++ Integra.Vision.Language/ASTNodes/QuerySections/OnConditionNode.cs
@@ -45,7 +45,7 @@ namespace Integra.Vision.Language.ASTNodes.QuerySections
             this.result = new PlanNode();
             this.result.Column = ChildrenNodes[0].Token.Location.Column;
             this.result.Line = ChildrenNodes[0].Token.Location.Line;
-            this.result.NodeType = (uint)PlanNodeTypeEnum.On;
+            this.result.NodeType = PlanNodeTypeEnum.On;
         }
 
         /// <summary>
