+++ Integra.Vision.Language/ASTNodes/QuerySections/SelectNode.cs
@@ -47,7 +47,7 @@ namespace Integra.Vision.Language.ASTNodes.QuerySections
             this.result = new PlanNode();
             this.result.Column = ChildrenNodes[0].Token.Location.Column;
             this.result.Line = ChildrenNodes[0].Token.Location.Line;
-            this.result.NodeType = (uint)PlanNodeTypeEnum.Select;
+            this.result.NodeType = PlanNodeTypeEnum.Select;
         }
 
         /// <summary>
@@ -78,7 +78,7 @@ namespace Integra.Vision.Language.ASTNodes.QuerySections
             foreach (var tupla in projection)
             {
                 PlanNode plan = new PlanNode();
-                plan.NodeType = (uint)PlanNodeTypeEnum.TupleProjection;
+                plan.NodeType = PlanNodeTypeEnum.TupleProjection;
                 plan.Children = new List<PlanNode>();
                 plan.Children.Add(tupla.Key);
                 plan.Children.Add(tupla.Value);
