+++ Integra.Vision.Language/ASTNodes/QuerySections/WithNode.cs
@@ -57,7 +57,7 @@ namespace Integra.Vision.Language.ASTNodes.QuerySections
             this.result = new PlanNode();
             this.result.Column = ChildrenNodes[0].Token.Location.Column;
             this.result.Line = ChildrenNodes[0].Token.Location.Line;
-            this.result.NodeType = (uint)PlanNodeTypeEnum.With;
+            this.result.NodeType = PlanNodeTypeEnum.With;
         }
 
         /// <summary>
